### 2020-09-10 #2656 @whedon
[PRE REVIEW]: DE-Sim: an object-oriented, discrete-event simulation tool for data-intensive modeling of complex systems in Python
**Submitting author:** @artgoldberg (<a href="http://orcid.org/0000-0003-2772-1484">Arthur Goldberg</a>)
**Repository:** <a href="https://github.com/KarrLab/de_sim" target ="_blank">https://github.com/KarrLab/de_sim</a>
**Version:** 0.1.1
