### 2020-09-10 #2656 @danielskatz
@artgoldberg - there are some problems to resolve before we move forward:

1. Your paper is too long - the tutorial part in particular should not be in the paper, but in the software repo documentation.  See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
2. There is no explicit "Statement of need" section.
3. There are some DOI problems - please make sure your .bib file only contains items that are cited in the paper, and that all DOIs are valid and in that files, where they exist.  Note that these DOI results from whedon are suggestions - they are not always correct or complete.

Please make changes to the paper (and bib), then add ```@whedon generate pdf``` as a new comment here to regenerate the PDF.  Feel free to iterate.  You can also use ```@whedon check references```.

I will mark this as paused for now.  Once you think you are ready to proceed, please tag @openjournals/joss-eics to let me (or the AEiC on duty) know.
