### 2020-09-10 214bdac
rename and reorder Jupyter notebooks; remove "pip install" commands from Jupyter notebooks; update joss_paper/de_sim_paper.pdf to version submitted to JOSS
