### 2020-09-12 #2656 @artgoldberg
Thank you @danielskatz I will address the issues you and whedon raise shortly.
Arthur
