### 2020-09-12 686b545
update paper as per JOSS editor
