### 2020-09-12 ac878f7
Remove articles in the bib that are not cited in the paper
