+++ joss_paper/paper.bib
@@ -26,26 +26,6 @@
   isbn={978-0136062127}
 }
 
-@article{dahl1966simula,
-  title={{SIMULA}: an {ALGOL}-based simulation language},
-  author={Dahl, Ole-Johan and Nygaard, Kristen},
-  journal={Communications of the ACM},
-  volume={9},
-  number={9},
-  pages={671--678},
-  year={1966},
-  doi={10.1145/365813.365819}
-}
-
-@incollection{nygaard1978development,
-  title={The development of the {SIMULA} languages},
-  author={Nygaard, Kristen and Dahl, Ole-Johan},
-  booktitle={{History of Programming Languages}},
-  pages={439--480},
-  year={1978},
-  doi={10.1145/800025.1198392}
-}
-
 @inproceedings{mueller2001simulation,
   title={The simulation semantics of {SystemC}},
   author={Mueller, Wolfgang and Ruf, Juergen and Hoffmann, Dirk and Gerlach, Joachim and Kropf, Thomas and Rosenstiehl, Wolfgang},
@@ -53,14 +33,6 @@
   pages={64--70},
   year={2001},
   doi={10.1109/DATE.2001.915002}
-}
-
-@article{ieee2012ieee,
-  title={{IEEE} Standard for standard {SystemC} language reference manual},
-  author={{IEEE Standards Association}},
-  journal={IEEE Computer Society},
-  year={2012},
-  doi={10.1109/IEEESTD.2012.6134619}
 }
 
 @inproceedings{concannon2003dynamic,
@@ -80,62 +52,12 @@
   year={2020}
 }
 
-@misc{karr2020wc_lang,
-  title = {{WC-Lang}: a language for composite, multi-algorithmic whole-cell models},
-  author={Karr, Jonathan R. and Goldberg, Arthur P. and Chew, Yin Hoon},
-  url = {https://github.com/KarrLab/wc_lang},
-  note = {Accessed: 2020-06-17},
-  year={2020}
-}
-
-@article{perumalla2012discrete,
-  title={Discrete event modeling and massively parallel execution of epidemic outbreak phenomena},
-  author={Perumalla, Kalyan S and Seal, Sudip K},
-  journal={Simulation},
-  volume={88},
-  number={7},
-  pages={768--783},
-  year={2012},
-  doi={10.1177/0037549711413001}
-}
-
 @misc{matloff2008introduction,
   title={Introduction to discrete-event simulation and the {SimPy} language},
   author={Matloff, Norm},
   year={2008},
   url={https://web.cs.ucdavis.edu/~matloff/matloff/public_html/156/PLN/DESimIntro.pdf},
   note = {Accessed: 2020-06-30},
-}
-
-@inproceedings{bahouth2007revisiting,
-  title={Revisiting the issue of performance enhancement of discrete event simulation software},
-  author={Bahouth, Alex and Crites, Steven and Matloff, Norman and Williamson, Todd},
-  booktitle={40th Annual simulation symposium (ANSS'07)},
-  pages={114--122},
-  year={2007},
-  organization={IEEE}
-}
-
-@article{Jefferson1985,
-  abstract = {Virtual time is a new paradigm for organizing and synchronizing distributed systems which can be applied to such problems as distributed discrete event simulation and distributed database concurrency control. Virtual time provides a flexible abstraction of real time in much the same way that virtual memory provides an abstraction of real memory. It is implemented using the Time Warp mechanism, a synchronization protocol distinguished by its reliance on lookahead-rollback, and by its implementation of rollback via antimessages.},
[…]
