### 2020-09-12 #2656 @artgoldberg
Hello openjournals/joss-eics:

The paper had included a section that contained the statement of need, but without that section title. I have renamed it "Statement of need".

I have renamed the tutorial from the paper. It is available on-line in a Jupyter notebook at https://sandbox.karrlab.org/notebooks/de_sim/1.%20DE-Sim%20tutorial.ipynb.

The article proof generated by whedon looks fine.

The missing and invalid DOIs still need to be fixed.

I recommend these four reviewers, all of whom are experts in discrete-event simulation. If you would like, I can also select a few potential reviewers from the list of people have already agreed to review for JOSS.

David Jefferson
Computer Scientist
Lawrence Livermore National Laboratory
drjefferson at Google's email
github: no account, I suspect

Christopher D. Carothers
Director, Center for Computational Innovations
Professor of Computer Science
Rensselaer Polytechnic Institute
chris.carothers at Google's email
github: carothersc

Richard Fujimoto
Regents' Professor
School of Computational Science and Engineering
Center for Experimental Research in Computer Systems (CERCS)
Georgia Institute of Technology
fujimoto [at] cc [dot] gatech [dot] edu
github: no account, I suspect

Robert Clayton Blake
Senior Software Engineer
Biochemical and Biophysical Systems Group
Lawrence Livermore National Laboratory
blake14 [at] llnl [dot] gov
github: rblake-llnl
