+++ joss_paper/paper.bib
@@ -36,12 +36,16 @@
 }
 
 @inproceedings{concannon2003dynamic,
+  author = {Concannon, Kieran H. and Hunter, Kim I. and Tremble, Jillian M.},
   title={Dynamic scheduling {II}: {SIMUL8}-planner simulation-based planning and scheduling},
-  author={Concannon, Kieran H and Hunter, Kim I and Tremble, Jillian M},
-  booktitle={{Proceedings of the 35th Conference on Winter simulation}},
-  pages={1488--1493},
-  year={2003},
-  doi={10.1145/1030818.1031019}
+  year = {2003},
+  isbn = {0780381327},
+  publisher = {Winter Simulation Conference},
+  pages = {1488–1493},
+  numpages = {6},
+  location = {New Orleans, Louisiana},
+  series = {WSC '03},
+  doi={10.1109/WSC.2003.1261534}
 }
 
 @misc{goldberg2020wc_sim,
@@ -170,41 +174,45 @@
 }
 
 @inproceedings{clune2006discrete,
-  title={Discrete event and hybrid system simulation with simevents},
+  title={{Discrete Event and Hybrid System Simulation with SimEvents}},
   author={Clune, Michael I and Mosterman, Pieter J and Cassandras, Christos G},
   booktitle={Proceedings of the 8th international workshop on discrete event systems},
   pages={386--387},
-  year={2006}
+  year={2006},
+  doi={10.1109/wodes.2006.382398}
 }
 
 @inproceedings{rice2005simscript,
-  title={The SIMSCRIPT III programming language for modular object-oriented simulation},
+  title={The {SIMSCRIPT III} programming language for modular object-oriented simulation},
   author={Rice, Stephen V and Marjanski, Ana and Markowitz, Harry M and Bailey, Stephen M},
   booktitle={Proceedings of the Winter Simulation Conference, 2005.},
   pages={10--pp},
   year={2005},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/wsc.2005.1574302}
 }
 
 @article{malengier2018odes,
-  title={ODES: a high level interface to ODE and DAE solvers},
+  title={ODES: a high level interface to {ODE} and {DAE} solvers},
   author={Malengier, Benny and Ki{\v{s}}on, Pavol and Tocknell, James and Abert, Claas and Bruckner, Florian and Bisotti, Marc-Antonio},
   journal={Journal of Open Source Software},
   volume={3},
   number={22},
   pages={165},
-  year={2018}
+  year={2018},
+  doi={10.21105/joss.00165}
 }
 
 @article{hart2011pyomo,
-  title={Pyomo: modeling and solving mathematical programs in Python},
+  title={Pyomo: modeling and solving mathematical programs in {Python}},
   author={Hart, William E and Watson, Jean-Paul and Woodruff, David L},
   journal={Mathematical Programming Computation},
   volume={3},
   number={3},
   pages={219},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12532-011-0026-8}
 }
 
 @article{karr2020objtables,
@@ -223,7 +231,8 @@
   number={3},
   pages={1331--1340},
   year={2002},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/s0006-3495(02)73903-9}
 }
 
 @article{hindmarsh2005sundials,
