### 2020-09-13 #2656 @artgoldberg
Hello @openjournals/joss-eics

I have fixed the missing and invalid DOIs, and a few problems with capitalization in the titles of references.

I believe that all of the issues Daniel S Katz raised have been addressed.

Thanks
Arthur
