### 2020-09-13 #2656 @danielskatz
The paper is still quite long, compared to the JOSS recommendation of 250-1000 words.  I suggest removing the case study and most of the availability section, and just pointing to the documentation and repository instead.
