### 2020-09-17 dd5d3e4
shorten paper as per request from JOSS editor: "I suggest removing the case study and most of the availability section"
