+++ joss_paper/paper.bib
@@ -70,7 +70,7 @@
   doi = {10.1109/PADS.2000.847144},
   journal = {Proceedings of the Fourteenth Workshop on Parallel and Distributed Simulation},
   pages = {53--60},
-  title = {{ROSS}: a high-performance, low memory, modular time warp system},
+  title = {{ROSS}: a high-performance, low memory, modular {Time Warp} system},
   url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=847144},
   volume = {62},
   year = {2000}
