### 2020-09-17 df72036
remove unneeded citations from bib
