+++ joss_paper/paper.bib
@@ -155,17 +155,6 @@
   doi={10.1016/j.cell.2012.05.044}
 }
 
-@article{gillespie1977exact,
-  title={Exact stochastic simulation of coupled chemical reactions},
-  author={Gillespie, Daniel T},
-  journal={Journal of Physical Chemistry},
-  volume={81},
-  number={25},
-  pages={2340--2361},
-  year={1977},
-  doi={10.1021/j100540a008}
-}
-
 @book{fishman2013discrete,
   title={Discrete-event simulation: modeling, programming, and analysis},
   author={Fishman, George S},
@@ -192,56 +181,10 @@
   doi={10.1109/wsc.2005.1574302}
 }
 
-@article{malengier2018odes,
-  title={ODES: a high level interface to {ODE} and {DAE} solvers},
-  author={Malengier, Benny and Ki{\v{s}}on, Pavol and Tocknell, James and Abert, Claas and Bruckner, Florian and Bisotti, Marc-Antonio},
-  journal={Journal of Open Source Software},
-  volume={3},
-  number={22},
-  pages={165},
-  year={2018},
-  doi={10.21105/joss.00165}
-}
-
-@article{hart2011pyomo,
-  title={Pyomo: modeling and solving mathematical programs in {Python}},
-  author={Hart, William E and Watson, Jean-Paul and Woodruff, David L},
-  journal={Mathematical Programming Computation},
-  volume={3},
-  number={3},
-  pages={219},
-  year={2011},
-  publisher={Springer},
-  doi={10.1007/s12532-011-0026-8}
-}
-
 @article{karr2020objtables,
   title={ObjTables: structured supplementary spreadsheets that promote data quality, reuse, and integration},
   author={Karr, Jonathan R and Liebermeister, Wolfram and Goldberg, Arthur P and Sekar, John A P and Shaikh, Bilal},
   journal={arXiv},
   page={2005.05227},
   year={2020}
-}
-
-@article{mahadevan2002dynamic,
-  title={Dynamic flux balance analysis of diauxic growth in Escherichia coli},
-  author={Mahadevan, Radhakrishnan and Edwards, Jeremy S and Doyle III, Francis J},
-  journal={Biophysical journal},
-  volume={83},
-  number={3},
-  pages={1331--1340},
-  year={2002},
-  publisher={Elsevier},
-  doi={10.1016/s0006-3495(02)73903-9}
-}
-
-@article{hindmarsh2005sundials,
-  title={{SUNDIALS}: Suite of nonlinear and differential/algebraic equation solvers},
-  author={Hindmarsh, Alan C and Brown, Peter N and Grant, Keith E and Lee, Steven L and Serban, Radu and Shumaker, Dan E and Woodward, Carol S},
-  journal={ACM Transactions on Mathematical Software (TOMS)},
-  volume={31},
-  number={3},
-  pages={363--396},
-  year={2005},
-  publisher={ACM}
 }
