### 2020-09-17 #2656 @artgoldberg
Hello @danielskatz

Since our paper is longer than the length recommended by JOSS I have reduced it from 2237 to 1856 words by removing the case study and most of the availability section as you recommended. (However, I also left a pointer to the Jupyter notebooks which are at a different website.)

I would like to add these reviewers to the list I previously recommended. They are likely to be more available.

Misbah Mubarak
Assistant computer scientist
Argonne National Laboratory
Mathematics and Computer Science Division
mmubarak [AT] anl.gov
github: no account, I suspect

Elsa J. Gonsiorowski
Software Developer
Lawrence Livermore National Laboratory
gonsie [AT] llnl.gov
github: gonsie

Gabriel Wainer
Professor
Department of Systems and Computer Engineering
Carleton University
Gabriel.Wainer [AT] sce.carleton.ca
github: gwainer

Thank you
Arthur
