### 2020-09-17 #2656 @danielskatz
👋 @rblake-llnl - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
