### 2020-09-18 #2656 @carothersc
Hi Dan - sure, I know the DE-Sim group pretty well. Happy to review this.

Regards,
Chris

On Thu, Sep 17, 2020 at 7:55 PM Daniel S. Katz <notifications@github.com>
wrote:

> 👋 @carothersc <https://github.com/carothersc> - would you be willing to
> review this submission for JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
