### 2020-09-23 KarrLab/de_sim#55 @gonsie
Failing Tests
This issue is related to openjournals/joss-reviews#2685

I’m running Python 3.8 on a Mac.  When running `python -m pytest tests` I am getting 2 warnings and 6 errors. Here is the summary output:

```
============================================ warnings summary ============================================
de_sim/simulation_object.py:473
/Users/gonsiorowski1/Projects/JOSS/de_sim/de_sim/simulation_object.py:473: UserWarning: SimulationObject 'Example' definition does not inherit or provide a non-empty 'messages_sent'.
      warnings.warn("SimulationObject '{}' definition does not inherit or provide a "

de_sim/simulation_object.py:473
/Users/gonsiorowski1/Projects/JOSS/de_sim/de_sim/simulation_object.py:473: UserWarning: SimulationObject 'SOwithDefaultClassPriority' definition does not inherit or provide a non-empty 'messages_sent'.
warnings.warn("SimulationObject '{}' definition does not inherit or provide a "

-- Docs: https://docs.pytest.org/en/stable/warnings.html
======================================== short test summary info =========================================
            ERROR tests/test_simulator.py
            ERROR tests/test_utilities.py
            ERROR tests/examples/test_jupyter_examples.py
            ERROR tests/examples/test_phold.py
            ERROR tests/examples/test_random_walk.py
            ERROR tests/examples/test_sirs.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 6 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
===================================== 2 warnings, 6 errors in 3.65s ======================================
```

It looks like most of errors come from:

```
ModuleNotFoundError: No module named 'capturer'
```
