### 2020-09-23 KarrLab/de_sim#55 @jonrkarr
Hi, the tests require additional dependencies beyond the package itself. These dependencies aren't installed by default because they aren't necessary for most users. These dependencies can be installed by running `pip install de_sim[tests]` or `pip install -r tests/requirements.txt` (if you have cloned the repository). Our CI system is setup to install these dependencies.
