### 2020-09-23 bf4b255
Adding instructions for installing dependencies of the tests
