### 2020-09-23 KarrLab/de_sim#55 @jonrkarr
I added these instructions to the [Guide to contributors](https://github.com/KarrLab/de_sim/blob/master/CONTRIBUTING.md).
