### 2020-09-23 KarrLab/de_sim#56 @gonsie
Warning when running examples
This issue is related to openjournals/joss-reviews#2685

When running an example I see the following warning printout:

```
> python -m de_sim.examples.phold 8 0.5 100
/.../homebrew/opt/python@3.8/bin/../Frameworks/Python.framework/Versions/3.8/lib/python3.8/codecs.py:905: RuntimeWarning: line buffering (buffering=1) isn't supported in binary mode, the default buffer size will be used
  file = builtins.open(filename, mode, buffering)
```
