### 2020-09-24 KarrLab/de_sim#55 @gonsie
I’m not able to run the tests successfully. I’m getting 1 failure and 1 hang.

The failure:
```
AssertionError: 'git@github.com:karrlab/de_sim' not found in ['https://github.com/karrlab/de_sim.git', 'git@github.com:karrlab/de_sim.git', 'ssh://git@github.com/karrlab/de_sim.git']

tests/test_simulation_metadata.py:64: AssertionError
```

And the 11th or 12th test of test_simulator.py is hanging.
