### 2020-09-24 7083cbf
accepting alternatively correct Git URLs
