### 2020-09-24 KarrLab/de_sim#55 @jonrkarr
I corrected the failing metadata test. The code is fine; the assertion is overly strict and was incorrectly failing in your environment. This is pushed to GitHub and PyPI.

@artgoldberg can you help debug the hanging test? It works fine in our environments (our machines, CircleCI, Mac, Linux) which use Python 3.7. I think its worth creating a Docker image for Python 3.8 to see if this is specific to Python 3.8.
