### 2020-09-25 d28f3cb
ensure that de_sim tests pass on Python 3.8; fix the path used by preprocess(); add ipykernel to the modules required by testing
