### 2020-09-25 07b1345
bump version to 0.1.4; stop issuing warning about missing messages_sent
