### 2020-09-25 KarrLab/de_sim#55 @artgoldberg
Hi @gonsie

I tested `de_sim` on the lastest micro-version of Python (3.8.5) in Ubuntu. I wasn't able to reproduce a hanging test, but the tests needed another module, `ipykernel`, which has been added to `tests/requirements.txt`.

In addition, the warning `UserWarning: SimulationObject ... does not inherit or provide a non-empty 'messages_sent'` has been removed.

I've pushed these updates to GitHub and PyPI. The new version of `de_sim` is `0.1.4`.

Thanks
Arthur
