### 2020-09-25 faa94dd
ignore warnings in codecs that are generated by logging2.FileHandler() calls; logging2 is an archived repo
