+++ de_sim/config/core.py
@@ -66,4 +66,12 @@ def get_debug_logs(extra=None):
         :obj:`wc_utils.debug_logs.core.DebugLogsManager`: a `DebugLogsManager`
     """
     config = get_debug_logs_config(extra)
-    return DebugLogsManager().setup_logs(config)
+    import warnings
+    with warnings.catch_warnings():
+        # Ignore RuntimeWarning warnings in codecs that are generated by logging2.FileHandler() calls
+        # logging2 is an archived repo
+        warnings.filterwarnings("ignore",
+                                category=RuntimeWarning,
+                                module='codecs')
+        manager = DebugLogsManager().setup_logs(config)
+    return manager
