### 2020-09-25 210d7c8
finish ignore warnings in codecs that are generated by logging2.FileHandler() calls
