### 2020-09-26 030e9ed
Revert "finish ignore warnings in codecs that are generated by logging2.FileHandler() calls"
This reverts commit 210d7c8b5894a0b287f40cc64d5df23992864865.
