+++ de_sim/examples/debug_logs.py
@@ -6,17 +6,9 @@
 :License: MIT
 """
 
-import warnings
-
 from .config import core
 from wc_utils.debug_logs.core import DebugLogsManager
 
 # setup debug logs
 config = core.get_debug_logs_config()
-with warnings.catch_warnings():
-    # Ignore RuntimeWarning warnings in codecs that are generated by logging2.FileHandler() calls
-    # logging2 is an archived repo
-    warnings.filterwarnings("ignore",
-                            category=RuntimeWarning,
-                            module='codecs')
-    logs = DebugLogsManager().setup_logs(config)
+logs = DebugLogsManager().setup_logs(config)
