### 2020-09-26 92ceb7b
Revert "ignore warnings in codecs that are generated by logging2.FileHandler() calls; logging2 is an archived repo"
This reverts commit faa94dd1fb8352004ed2b1521dede6b0ff34a07d.
