+++ tests/test_utilities.py
@@ -119,29 +119,20 @@ class TestFastLogger(unittest.TestCase):
             self.assertTrue(capturer.get_text().endswith(message))
 
     def test_config(self):
-        import warnings
-        with warnings.catch_warnings():
-            # Ignore warnings in codecs that are generated by logging2.FileHandler() calls
-            # logging2 is an archived repo
-            warnings.filterwarnings("ignore",
-                                    category=RuntimeWarning,
-                                    module='codecs')
+        debug_config = core.get_debug_logs_config(cfg_path=('tests', 'fixtures/config/debug.default.cfg'))
+        debug_log_manager = DebugLogsManager().setup_logs(debug_config)
 
-            debug_config = core.get_debug_logs_config(cfg_path=('tests',
-                                                                 'fixtures/config/debug.default.cfg'))
-            debug_log_manager = DebugLogsManager().setup_logs(debug_config)
+        file_logger = debug_log_manager.logs['de_sim.debug.testing.file']
+        self.assertFalse(FastLogger(file_logger, 'debug').active)
+        self.assertTrue(FastLogger(file_logger, 'info').active)
 
-            file_logger = debug_log_manager.logs['de_sim.debug.testing.file']
-            self.assertFalse(FastLogger(file_logger, 'debug').active)
-            self.assertTrue(FastLogger(file_logger, 'info').active)
+        console_logger = debug_log_manager.logs['de_sim.debug.testing.console']
+        self.assertFalse(FastLogger(console_logger, 'debug').active)
+        self.assertFalse(FastLogger(console_logger, 'info').active)
+        self.assertTrue(FastLogger(console_logger, 'warning').active)
 
-            console_logger = debug_log_manager.logs['de_sim.debug.testing.console']
-            self.assertFalse(FastLogger(console_logger, 'debug').active)
-            self.assertFalse(FastLogger(console_logger, 'info').active)
-            self.assertTrue(FastLogger(console_logger, 'warning').active)
+        debug_config = core.get_debug_logs_config()
+        debug_log_manager = DebugLogsManager().setup_logs(debug_config)
 
-            debug_config = core.get_debug_logs_config()
-            debug_log_manager = DebugLogsManager().setup_logs(debug_config)
-
-            file_logger = debug_log_manager.logs['de_sim.debug.file']
-            self.assertFalse(FastLogger(file_logger, 'debug').active)
+        file_logger = debug_log_manager.logs['de_sim.debug.file']
+        self.assertFalse(FastLogger(file_logger, 'debug').active)
