### 2020-09-26 KarrLab/de_sim#56 @artgoldberg
Thanks for notifying us. The warning is generated by a logging package that's outside our control. I have disabled the warning notification by ignoring it in a lower code layer.

Regards
Arthur
