### 2020-09-28 KarrLab/de_sim#55 @artgoldberg
Hi @gonsie

Could you please try the unit tests again? I was not able to reproduce the hanging test you report.
If you can reproduce it, could you please report the specific MacOS and Python versions?

Thanks
Arthur
