### 2020-09-28 5b763d7
properly report failures
