### 2020-09-28 882cbda
better failure reporting
