### 2020-09-28 4c4026c
reset TIMEOUT to normal value
