### 2020-09-28 80cc409
add good error reporting
