### 2020-09-28 21f094e
improve error reporting
