+++ tests/joss_paper/test_gen_phold_space_time_plot.py
@@ -14,17 +14,18 @@ class TestPlotLog(unittest.TestCase):
     def test(self):
         # this tests two things:
         # 1) gen_phold_space_time_plot.py
-        # 2) the use of self.fast_plot_file_logger in de_sim/de_sim/simulation_object.py, which cannot be conveniently
-        # unit-tested because doing so requires that config state be changed before de_sim.plot.file logger is created
-        # we run gen_phold_space_time_plot.py in a separate process to execute that code
-        # also see comment LoggerConfigurator().from_dict() regarding shared loggers
+        # 2) the use of self.fast_plot_file_logger in de_sim/de_sim/simulation_object.py, which cannot
+        # be conveniently unit-tested because doing so requires that config state be changed before
+        # de_sim.plot.file logger is created we run gen_phold_space_time_plot.py in a separate process
+        # to execute that code.
+        # also see the comment in LoggerConfigurator().from_dict() regarding shared loggers.
         try:
             command = ["python3", "joss_paper/gen_phold_space_time_plot.py"]
             result = subprocess.run(command,
                                     check=True,
                                     capture_output=True)
-        except CalledProcessError as e:
-            e_msg = f"""'{" ".join(command)}' failed: {e.stderr}"""
+        except subprocess.CalledProcessError as e:
+            e_msg = f"""'{" ".join(command)}' failed:\n{e.stderr.decode("utf-8")}"""
             self.fail(e_msg)
 
         self.assertIn('space-time diagram written', result.stdout.decode("utf-8"))
