### 2020-09-28 KarrLab/de_sim#55 @artgoldberg
Hi @gonsie

Alternatively, to confirm that DE Sim works under Python 3.8, you could run its unit tests in a Docker container running Python 3.8:

Docker commands (requires Docker):

    docker pull python:3.8.6
    # run an interactive shell in the Docker container
    docker run --interactive --tty python:3.8.6  bash

Shell commands to execute in the running Docker container:

    git clone https://github.com/KarrLab/de_sim.git
    cd de_sim
    # obtain all of DE Sim's dependencies from PyPI:
    pip install -r requirements.txt -r tests/requirements.txt
    pip install pytest
    # setup the bash environment for subprocesses
    export PYTHONPATH="$(pwd)"
    pytest tests/

You should get this output:

    tests/test_api.py .                                         [  1%]
    tests/test_checkpoint.py ....                               [  5%]
    tests/test_errors.py .                                      [  6%]
    tests/test_event.py ...                                     [  9%]
    tests/test_event_message.py ....                            [ 13%]
    tests/test_init.py .                                        [ 14%]
    tests/test_mock_simulation_object.py .                      [ 15%]
    tests/test_simulation_checkpoint_object.py ...              [ 18%]
    tests/test_simulation_config.py ........                    [ 26%]
    tests/test_simulation_metadata.py ........                  [ 34%]
    tests/test_simulation_object.py .............               [ 47%]
    tests/test_simulator.py ..................s..........       [ 76%]
    tests/test_template_sim_objs.py ...                         [ 79%]
    tests/test_utilities.py .......                             [ 86%]
    tests/test_visualize.py ..                                  [ 88%]
    tests/config/test_core.py ..                                [ 90%]
    tests/examples/test_jupyter_examples.py .                   [ 91%]
    tests/examples/test_minimal_simulation.py ..                [ 93%]
    tests/examples/test_phold.py ...                            [ 96%]
    tests/examples/test_random_walk.py .                        [ 97%]
    tests/examples/test_sirs.py ..                              [ 99%]
    tests/joss_paper/test_gen_phold_space_time_plot.py .        [100%]

    ============ 99 passed, 1 skipped in 192.93s (0:03:12) ===========

The tests take under 5 min. to run on my laptop, which has a 2.9 GHz Intel Core i5 with 16 GB RAM. They take a while because `test_jupyter_examples.py` runs the DE Sim performance tests.

Arthur
