### 2020-09-29 KarrLab/de_sim#55 @artgoldberg
Thanks @gonsie. `capturer.CaptureOutput()` is failing to initialize here. If you share your computing environment, such as the versions of OSX, virtualenv, and Python you're using I can report this problem as an issue for `capturer`.

With respect to DE Sim, to confirm that it passes all tests under Python 3.8, you could run its unit tests in a Docker container, as I suggested yesterday.

Regards
Arthur
