### 2020-09-29 #2685 @artgoldberg
Hi @danielskatz. @gonsie finds that one of DE Sim's unit tests fails because of a problem in the [`capturer`](https://github.com/xolox/python-capturer) testing library. The test passes on multiple platforms (local, Docker on local, Docker in CircleCI) for me and I have not reproduced the `capturer` failure thus far. See [`KarrLab/de_sim#55`](https://github.com/KarrLab/de_sim/issues/55) for the details.

How would you suggest we proceed?

Thanks
Arthur
