### 2020-09-30 bb266a9
capture stdout using contextlib.redirect_stdout instead of CaptureOutput, because CaptureOutput fails in test run by JOSS paper reviewer; see https://github.com/KarrLab/de_sim/issues/55; bump version
