### 2020-09-30 KarrLab/de_sim#55 @artgoldberg
Hi @gonsie

I have replaced `capturer.CaptureOutput()` with another method in the code that failed for you. A new version (0.1.7) of DE Sim that contains this change is available in the [DE Sim repo on GitHub](https://github.com/KarrLab/de_sim) and in the [de-sim package on PyPI](https://pypi.org/project/de-sim/).

I don't know whether this change will fix the failure, as I have not reproduced the problem. Could you please try it?

In addition, I've filed an issue with [`capturer`](https://github.com/xolox/python-capturer) regarding the failure you experienced. If you wish to provide more details regarding your computing environment I will add them to the issue.

Thanks
Arthur
