### 2020-09-30 #2685 @artgoldberg
Thanks @danielskatz

OK. I have replaced the failing code in DE Sim's tests with a method from Python's standard library that will hopefully avoid the problem. See [`KarrLab/de_sim#55`](https://github.com/KarrLab/de_sim/issues/55) for further details.

Arthur
