### 2020-10-16 af74951
make joss_paper.md with png files to see if it renders on GitHub; clean out old/
