### 2020-10-17 da4c958
replace pdf figures with png ones so they render in github markdown; organize figures in a folder; rm test joss_paper.md
