### 2020-10-19 #2685 @carothersc
Hi all - given that Elsa has confirmed running the software on her side do I need to do the same ? Also, can someone confirm that "pip" won't do anything bad to local "conda" repos?

Thanks everyone!,
Chris
