### 2020-10-19 #2685 @danielskatz
We expect the reviewers to be independent in terms of investigation, though they may work together in documenting problems and suggesting solutions.
