### 2020-10-19 #2685 @danielskatz
Re pip, I suggest the safest thing is to create a new conda environment for this review
