### 2020-10-20 #2685 @artgoldberg
@carothersc another approach that would not risk your Python computing environment would be to run DE Sim's tests in a Docker image, as I describe in [this comment](https://github.com/KarrLab/de_sim/issues/55#issuecomment-700329431).
