### 2020-10-20 KarrLab/de_sim#57 @gonsie
Dead links in tutorial 1
This issue is related to openjournals/joss-reviews#2685

Tutorial `1. DE-Sim tutorial.ipynb` has dead links. It references an “Advanced DE Sim notebook” which does not exist.
