### 2020-10-21 #2685 @gonsie
I have completed my review and I approve this paper for acceptance.

I do think there is room for improvement with the model examples given. They are given as complete models, with no description of the process for creating a model. There are a lot of python features that model developers can take advantage of, but it may be hard to discover these without direct support from the de_sim team. Other than that, everything looks great.
