### 2020-10-21 #2685 @artgoldberg
Thanks @gonsie . Can DE-Sim [issue #55](https://github.com/KarrLab/de_sim/issues/55) be closed?

Arthur
