### 2020-10-21 cf63d72
instead of requiring that a simulation's output dir be empty, just require that it contains no checkpoints
