### 2020-10-21 #2685 @danielskatz
Thanks @gonsie - I'm a little confused, since you both say "I approve this paper for acceptance" and you have checked all the review boxes, but you also have opened issues. Does this mean that these issues are optional to you: suggestions but not requirements?
