### 2020-10-21 #2685 @gonsie
@danielskatz yes. The issues are optional. They are small issues that need not be fixed for my acceptance.
