### 2020-10-26 64be7eb
Restructure EventMessages to look like dataclasses. Attributes are defined using standard PEP 526 type annotations. Save default values and type declarations for future use.
