+++ tests/test_event.py
@@ -91,7 +91,8 @@ class TestEvent(unittest.TestCase):
 
         class TestMsg(de_sim.EventMessage):
             'docstring'
-            msg_field_names = ['attr1', 'attr2']
+            attr1: str
+            attr2: str
         vals = ['att1_val', 'att2_val']
         test_msg = TestMsg(*vals)
         times = (0, 1)
@@ -126,7 +127,7 @@ class TestEvent(unittest.TestCase):
         # self.assertEqual(data+vals, ev.render(as_list=True))
 
         class NoBodyMessage(de_sim.EventMessage):
-            """A message with no `msg_field_names`"""
+            "A message with no attributes"
         ev2 = de_sim.Event(0, 1, ExampleSimulationObject('sender'), ExampleSimulationObject('receiver'),
                     NoBodyMessage())
         self.assertIn('\t'.join(de_sim.Event.BASE_HEADERS), ev2.custom_header())
