### 2020-10-27 #2685 @artgoldberg
Hello @danielskatz

If Chris is not able to review our paper, perhaps another reviewer could step in. Among the other potential reviewers I recommended in the PRE REVIEW, I would suggest Robert Clayton Blake. He has the appropriate discrete-event simulation and software expertise, and has, in general, been responsive in recent years.

Regards
Arthur
