### 2020-10-27 3e54822
Update the docstring for EventMessage to describe its improved syntax and semantics based on PEP 526 and Python's Data Classes.
