### 2020-11-01 KarrLab/de_sim#63 @yadudoc
`test_gen_phold_space_time_plot.py` is failing
Hi,

This issue is related to openjournals/joss-reviews#2685

I'm running tests on an Ubuntu 20.04, with python3.7 in a clean conda env, and I'm finding the installation, setup and tests
to be all nicely documented and in working order. This is the only test that I'm seeing fail on my machine: `tests/joss_paper/test_gen_phold_space_time_plot.py`

I've repeated the test with just this one a few times, and it errors out consistently.

Here's the traceback:
```
            retcode = process.poll()
            if check and retcode:
                raise CalledProcessError(retcode, process.args,
>                                        output=stdout, stderr=stderr)
E               subprocess.CalledProcessError: Command '['python3', 'joss_paper/figures/gen_phold_space_time_plot.py']' returned non-zero exit status 1.

../../anaconda3/envs/de_sim/lib/python3.7/subprocess.py:512: CalledProcessError

self = <test_gen_phold_space_time_plot.TestPlotLog testMethod=test>

    def test(self):
        # this tests two things:
        # 1) gen_phold_space_time_plot.py
        # 2) the use of self.fast_plot_file_logger in de_sim/de_sim/simulation_object.py, which cannot
        # be conveniently unit-tested because doing so requires that config state be changed before
        # de_sim.plot.file logger is created we run gen_phold_space_time_plot.py in a separate process
        # to execute that code.
        # also see the comment in LoggerConfigurator().from_dict() regarding shared loggers.
        try:
            command = ["python3", "joss_paper/figures/gen_phold_space_time_plot.py"]
            result = subprocess.run(command,
                                    check=True,
                                    capture_output=True)
        except subprocess.CalledProcessError as e:
            e_msg = f"""'{" ".join(command)}' failed:\n{e.stderr.decode("utf-8")}"""
>           self.fail(e_msg)
E           AssertionError: 'python3 joss_paper/figures/gen_phold_space_time_plot.py' failed:
E           Traceback (most recent call last):
E             File "joss_paper/figures/gen_phold_space_time_plot.py", line 48, in <module>
E               create_phold_space_time_diagram()
E             File "joss_paper/figures/gen_phold_space_time_plot.py", line 38, in create_phold_space_time_diagram
E               run_phold(8)
E             File "joss_paper/figures/gen_phold_space_time_plot.py", line 29, in run_phold
E               RunPhold.main(args)
E             File "/home/yadu/anaconda3/envs/de_sim/lib/python3.7/site-packages/de_sim/examples/phold.py", line 135, in main
E               event_num = simulator.simulate(args.max_time).num_events
E             File "/home/yadu/anaconda3/envs/de_sim/lib/python3.7/site-packages/de_sim/simulator.py", line 583, in simulate
E               self._simulate()
E             File "/home/yadu/anaconda3/envs/de_sim/lib/python3.7/site-packages/de_sim/simulator.py", line 688, in _simulate
E               next_sim_obj._BaseSimulationObject__handle_event_list(next_events)
E             File "/home/yadu/anaconda3/envs/de_sim/lib/python3.7/site-packages/de_sim/simulation_object.py", line 276, in __handle_event_list
E               handler(self, event)
E             File "/home/yadu/anaconda3/envs/de_sim/lib/python3.7/site-packages/de_sim/examples/phold.py", line 57, in handle_simulation_event
E               self.log_debug_msg("{:8.3f}: {} sending to self".format(self.time, self.name))
E             File "/home/yadu/anaconda3/envs/de_sim/lib/python3.7/site-packages/de_sim/examples/phold.py", line 76, in log_debug_msg
E               log = logs.get_log('de_sim.debug.example.console')
E             File "/home/yadu/anaconda3/envs/de_sim/lib/python3.7/site-packages/wc_utils/debug_logs/core.py", line 69, in get_log
E               list(logs.keys())))
E           ValueError: log named 'de_sim.debug.example.console' not found in logs '[]'.

tests/joss_paper/test_gen_phold_space_time_plot.py:29: AssertionError
============================================================================================= short test summary info =============================================================================================
FAILED tests/joss_paper/test_gen_phold_space_time_plot.py::TestPlotLog::test - AssertionError: 'python3 joss_paper/figures/gen_phold_space_time_plot.py' failed:
```
