### 2020-11-01 KarrLab/de_sim#64 @yadudoc
Some tests open up new windows to display plots
Hi,

This issue is related to openjournals/joss-reviews#2685

I notice that some tests pop up new windows to display plots. Do you think it might be better to write the plots to files and check file size to confirm? I have not taken a peek at your CI system or checked for test harnesses so I might be missing any checks you might already have in place.
