### 2020-11-02 KarrLab/de_sim#65 @yadudoc
Add Pandas examples for Jupyter notebooks?
Hi,

This issue is related to openjournals/joss-reviews#2685

Let me start off by noting that I've had a quick walkthrough of your notebooks, and they are nicely organized, clear, and understandable. All of the notebooks worked for me on your hosted setup. In the JOSS writeup, you mention integrating with Pandas/Scipy and other existing data science tools as a key advantage. The second notebook on epidemic modeling is the only one that shows using pandas, and that too only for plotting. It might be useful to update the example to show how a tighter integration might look like.
