### 2020-11-02 #2685 @yadudoc
@artgoldberg, I have completed my review, I'm posting my notes below.

The `DE-Sim` repo appears to be in good order. It is well documented, and with the contributing guide, I was able to install, run tests and do some tinkering on my laptop without issues. The one error I found, is listed as a github issue.

The manuscript is written well and clearly lays out the problem space, the motivation, and the advantages offered by `DE-Sim`. The Jupyter notebooks complement the documentation well and serve well to explain how the software could be used. I have added a recommendation to expand on tool integration, a key feature described in the manuscript.

Required issues opened:
- [ ]  https://github.com/KarrLab/de_sim/issues/63

Optional issues opened:
- [ ] https://github.com/KarrLab/de_sim/issues/64
- [ ] https://github.com/KarrLab/de_sim/issues/65
