### 2020-11-02 #2685 @yadudoc
@danielskatz That issue(https://github.com/KarrLab/de_sim/issues/63) is the only blocker from my end. Apart from this one item, I would recommend accepting this paper.
