### 2020-11-04 #2685 @artgoldberg
I appreciate your feedback @yadudoc. Please excuse my slow response. I've been out campaigning in PA. I'll investigate your issues, especially #63 soon.

Arthur
