### 2020-11-05 KarrLab/de_sim#63 @artgoldberg
Hi @yadudoc

These steps reproduce this failure:

1. `pip install de-sim`
2. `git clone https://github.com/KarrLab/de_sim.git` to obtain the DE-Sim's tests
3. `cd de_sim`
4. `pytest tests/joss_paper/test_gen_phold_space_time_plot.py`, or another method to run the tests in `test_gen_phold_space_time_plot.py`

The problem with this approach is that two different installations of DE Sim are being used, first, the one made by `pip` in `/usr/local/lib/python3.7/site-packages`, and second, the one that was obtained by `git clone` or some other approach that obtains all of the unit test and their dependencies.

You could verify that the test in #4 passes by updating the PYTHONPATH environment variable to include the current directory with `PYTHONPATH="$PYTHONPATH:."` before running the test. Then only the DE Sim installation created by `git clone` will be used.

Arthur
