### 2020-11-05 KarrLab/de_sim#63 @yadudoc
Adding the repository root to `PYTHONPATH` worked, nice! I recommend adding this as a note to the contributing guide. With that, I'll close this issue.
