### 2020-11-05 #2685 @yadudoc
Hi @danielskatz, the issue (https://github.com/KarrLab/de_sim/issues/63) with the tests is now resolved. I recommend this paper for acceptance.
