### 2020-11-05 #2685 @danielskatz
👋 @artgoldberg - At this point I'm going to proofread the paper
