### 2020-11-05 KarrLab/de_sim#63 @artgoldberg
Thanks @yadudoc . I've update `CONTRIBUTING.md` as you suggest.
