### 2020-11-05 #2685 @artgoldberg
Thanks @danielskatz
I will do that. I'm also cleaning up a few things in the repo, such as making sure that all version numbers and publication dates are correct, and addressing a suggestions @yadudoc made in [issue #63](https://github.com/KarrLab/de_sim/issues/63).
