### 2020-11-05 1040cc0
Update CONTRIBUTING.md as requested by reviewer; Clean up several metadata files
