+++ joss_paper/README.txt
@@ -5,4 +5,4 @@ At the time of publication, it was compiled into pdf by using the Whedon paper p
 
 A.P.G. conceptualized, designed, and implemented DE-Sim.
 J.R.K. provided motivation and input for DE-Sim and helped test, document, and distribute DE-Sim.
-Both authors wrote and approved this manuscript.
+Both authors wrote and approved this paper.
