### 2020-11-06 03ea64d
As per the Semantic Versioning Specification, update version # to 1.0.0
