### 2020-11-06 #2685 @artgoldberg
Hi @danielskatz
What publication date should I put in the `paper.md` file?
