### 2020-11-06 d94a3e9
update trove classifiers
