### 2020-11-06 8733c6b
add publication (written on, as the JOSS editor says) dates
