+++ joss_paper/paper.md
@@ -19,7 +19,7 @@ authors:
 affiliations:
  - name: Icahn Institute for Data Science and Genomic Technology and Department of Genetics and Genomic Sciences, Icahn School of Medicine at Mount Sinai, New York, NY 10029, USA
    index: 1
-date: 18 August 2020
+date: 6 November 2020
 bibliography: paper.bib
 ---
 
