### 2020-11-06 d792d57
Proofread JOSS paper; Correct features list in README.md; Add URL to ObjTables citation
