+++ joss_paper/paper.bib
@@ -186,5 +186,6 @@
   author={Karr, Jonathan R and Liebermeister, Wolfram and Goldberg, Arthur P and Sekar, John A P and Shaikh, Bilal},
   journal={arXiv},
   page={2005.05227},
+  url = {https://arxiv.org/abs/2005.05227},
   year={2020}
 }
