### 2020-11-07 460a9c1
update citations to the JOSS paper that's about to be published
