### 2020-11-07 8bd6202
Update citations to JOSS paper.
