+++ docs/comparison.rst
@@ -1,4 +1,4 @@
 Comparison to other DES tools
 =============================
 
-Please see the `DE-Sim article <https://github.com/KarrLab/de_sim/blob/master/joss_paper/de_sim_paper.pdf>`_ (in preparation) for a comparison of *DE-Sim* with other DES tools.
+Please see Arthur P. Goldberg & Jonathan Karr. (2020). `DE-Sim: an object-oriented, discrete-event simulation tool for data-intensive modeling of complex systems in Python. Journal of Open Source Software, 5(55), 2685. <https://doi.org/10.21105/joss.02685>`_ for a comparison of *DE-Sim* with other DES tools.
