+++ joss_paper/README.txt
@@ -1,4 +1,5 @@
-This directory contains a paper published by JOSS.
+This directory contains a paper published by JOSS, Arthur P. Goldberg & Jonathan Karr. (2020). DE-Sim: an object-oriented, discrete-event simulation tool for data-intensive modeling of complex systems in Python. Journal of Open Source Software, 5(55), 2685, https://doi.org/10.21105/joss.02685.
+
 At the time of publication, it was compiled into pdf by using the Whedon paper preview service at https://whedon.theoj.org/.
 
 # Contributions
