### 2020-11-07 c5d79f9
clean up a couple of nits
