### 2020-11-09 #2685 @artgoldberg
@danielskatz FYI, it may take a few days for me to provide the final version of the paper and the archive DOI. I want my co-author Prof. Karr to have an opportunity to review them, but he's busy with a new child.
