### 2020-11-09 a79e29f
avoid PyLint W0613 warnings, unused-argument
