### 2020-11-10 78c9dd9
stop matplotlib.pyplot.show() from opening new windows during testing
