+++ tests/examples/test_random_walk.py
@@ -11,15 +11,17 @@ from argparse import Namespace
 from capturer import CaptureOutput
 
 from de_sim.examples.random_walk import RunRandomWalkSimulation
+from de_sim.testing.utilities_for_testing import unset_env_var
 
 
-class TestRandomStateVariableSimulation(unittest.TestCase):
+class TestRunRandomWalkSimulation(unittest.TestCase):
 
     def test_random_walk_simulation(self):
         with CaptureOutput(relay=False):
             with tempfile.NamedTemporaryFile() as file_like_obj:
-                out_file = file_like_obj.name
-                args = Namespace(max_time=10, plot_file=out_file, output=False)
-                self.assertTrue(0 < RunRandomWalkSimulation.main(args))
-                args = Namespace(max_time=10, plot_file=out_file, output=True)
-                self.assertTrue(0 < RunRandomWalkSimulation.main(args))
+                with unset_env_var('DISPLAY'):
+                    out_file = file_like_obj.name
+                    args = Namespace(max_time=10, plot_file=out_file, output=False)
+                    self.assertTrue(0 < RunRandomWalkSimulation.main(args))
+                    args = Namespace(max_time=10, plot_file=out_file, output=True)
+                    self.assertTrue(0 < RunRandomWalkSimulation.main(args))
