+++ de_sim/testing/utilities_for_testing.py
@@ -5,7 +5,8 @@
 :Copyright: 2018-2020, Karr Lab
 :License: MIT
 """
-
+import contextlib
+import os
 
 def make_args(args_dict, required, options):
     """ Make command line argument list, for testing argument parsers
@@ -26,3 +27,22 @@ def make_args(args_dict, required, options):
     for arg in required:
         args.extend([str(args_dict[arg])])
     return args
+
+
+@contextlib.contextmanager
+def unset_env_var(env_var):
+    """ Temporarily unset an environment variable
+
+    Args:
+        env_var (:obj:`str`): environment variable to unset
+    """
+    saved_env_var = None
+    if env_var in os.environ:
+        saved_env_var = os.environ[env_var]
+        del os.environ[env_var]
+
+    try:
+        yield
+    finally:
+        if saved_env_var is not None:
+            os.environ[env_var] = saved_env_var
