### 2020-11-14 a12f8d0
fix a couple of bib entries; insert about 10 edits each by me and Jonathan
