+++ joss_paper/paper.bib
@@ -107,7 +107,7 @@
 
 @inproceedings{mckinney2010data,
   title={Data structures for statistical computing in {Python}},
-  author={McKinney, Wes and others},
+  author={McKinney, Wes},
   booktitle={{Proceedings of the 9th Python in Science Conference}},
   volume={445},
   pages={51--56},
@@ -159,7 +159,8 @@
   title={Discrete-event simulation: modeling, programming, and analysis},
   author={Fishman, George S},
   year={2013},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  isbn={978-0-387-95160-7}
 }
 
 @inproceedings{clune2006discrete,
