### 2020-11-14 9af490d
a few final paper revisions
