### 2020-11-14 9d739be
write final publication date; clarify Simula description
