### 2020-11-15 #2685 @artgoldberg
Hi @danielskatz

The paper you compiled with Whedon looks fine. Jonathan and I made a couple of dozen small changes to fix typos, improve wording, and clarify a couple of points.

I've created a tagged release of de-sim, with [tag 1.0.4](https://github.com/KarrLab/de_sim/releases/tag/1.0.5).
This release is archived at [Zenodo](https://zenodo.org/record/4274852) with DOI 10.5281/zenodo.4274852.

Thanks

Arthur
