### 2020-11-16 #2685 @danielskatz
@openjournals/dev - note that in the xml to be submitted to Crossref, the ```~``` in the Matloff unstructured citation has been turned into a space.  There's probably a bug in parsing somewhere...  Can we (you) manually fix this in this case?  If so, now or after I do the actual accept?
