### 2020-11-16 #2685 @arfon
>@openjournals/dev - note that in the xml to be submitted to Crossref, the ~ in the Matloff unstructured citation has been turned into a space. There's probably a bug in parsing somewhere... Can we (you) manually fix this in this case? If so, now or after I do the actual accept?

@danielskatz - Crossref deposit XML is fixed up now and I've opened an issue (https://github.com/openjournals/whedon-api/issues/122) to track this bug.
