### 2020-11-16 #2685 @artgoldberg
@danielskatz Greatly appreciate your patience, effort, persistence and care reviewing our paper. Great catch of the `s/~/ /` bug in, it appears, the LaTeX converter.

I see that you enjoy sailing. Me too. If you come to New York during warm weather, let me know and I'll gladly take you sailing at the [City Island Yacht Club](http://www.cityislandyc.org/).
